Revert "Changes for QDK pipeline (#70)" (#80)
Reverting this change since it does not play nicely with our internal build. 
My apologies for the reversion - let's see how to best handle the cases that pass the local build on the public repo but fail as part of our internal pipeline in the future (I think we need a more consistent setup, or revise how to handle this).