Explicitly import QSharp.Core package for e2e build work.
The ongoing refactoring in qsharp-runtime means that soon Microsoft.Quantum.Simulators will not directly include the Microsoft.Quantum.QSharp.Core package, so any callers that expected that need to include that additional reference. Note that this only affects projects that either don't use the QDK or suppress the automatic open of the core packages via `IncludeQsharpCorePackages` set to false.

This is safe to do even before the refactor is complete because in the current system it's just a redundant reference.