Add extra nuget packages to iqsharp-base container
The %azure.target command currently has poor performance:

Looking at the cell duration telemetry, the 95th percentile for %azure.target commands is 50 seconds, and the 99.5 percentile is 2 min 43 seconds.

After the first time it is executed, the cell has sub-second performance (even after a kernel restart). It looks like this is due to the packages that need to be downloaded or unpacked from Nuget. Executing the %azure.target command for our three providers adds 55 nuget packages to the container.

By installing them in the base image first, we can improve the performance of this command, so that it is sub-second even for the first time it is run (assuming the user has not explicitly specified a nuget version not in our nuget cache).